Refactor, added Interface for Card, and implement French cards
