+++ src/main/java/com/test/vp/card/FrenchCard.java
@@ -1,0 +1,51 @@
+package com.test.vp.card;
+
+public class FrenchCard implements Card {
+
+	private Value value;
+	private Suit suit;
+
+	public FrenchCard(Value value, Suit suit) {
+		super();
+		this.value = value;
+		this.suit = suit;
+	}
+
+	public Value getValue() {
+		return value;
+	}
+
+	public Suit getSuit() {
+		return suit;
+	}
+
+	public void setValue(Value value) {
+		this.value = value;
+	}
+
+	public void setSuit(Suit suit) {
+		this.suit = suit;
+	}
+
+	@Override
+	public String toString() {
+		return "Card [value=" + value + ", suit=" + suit + "]";
+	}
+
+	@Override
+	public boolean equals(Object obj) {
+		if (this == obj)
+			return true;
+		if (obj == null)
+			return false;
+		if (getClass() != obj.getClass())
+			return false;
+		FrenchCard other = (FrenchCard) obj;
+		if (suit != other.suit)
+			return false;
+		if (value != other.value)
+			return false;
+		return true;
+	}
+
+}
